perf: change usage of f64 to f32

+++ filters/src/kernel.rs
@@ -2,8 +2,8 @@ use ndarray::prelude::*;
 use ndarray::Array;
 use std::iter::FromIterator;
 
-pub fn box_blur_kernel_1d(radius: usize) -> (Array2<f64>, Array2<f64>) {
-    let kernel = Array::<f64, _>::ones((1, radius * 2 + 1).into_shape());
+pub fn box_blur_kernel_1d(radius: usize) -> (Array2<f32>, Array2<f32>) {
+    let kernel = Array::<f32, _>::ones((1, radius * 2 + 1).into_shape());
 
     // Return normalized kernels
     let kernel_x = &kernel / kernel.sum();
@@ -12,20 +12,20 @@ pub fn box_blur_kernel_1d(radius: usize) -> (Array2<f64>, Array2<f64>) {
     (kernel_x, kernel_y)
 }
 
-pub fn box_blur_kernel_2d(radius: usize) -> Array2<f64> {
-    let kernel = Array::<f64, _>::ones((radius * 2 + 1, radius * 2 + 1).into_shape());
+pub fn box_blur_kernel_2d(radius: usize) -> Array2<f32> {
+    let kernel = Array::<f32, _>::ones((radius * 2 + 1, radius * 2 + 1).into_shape());
 
     // Return normalized kernel
     &kernel / kernel.sum()
 }
 
-pub fn gaussian_blur_kernel_1d(sigma: f64) -> (Array2<f64>, Array2<f64>) {
+pub fn gaussian_blur_kernel_1d(sigma: f32) -> (Array2<f32>, Array2<f32>) {
     assert!(sigma > 0.0, "--sigma should be > 0.0");
 
     // Generate a 1×N Gaussian kernel
-    let radius = sigma.ceil() as i64 * 3;
+    let radius = sigma.ceil() as i32 * 3;
     let kernel = Array::from_iter(
-        (-radius..=radius).map(|x| (-(x.pow(2) as f64) / (2.0 * sigma.powi(2))).exp()),
+        (-radius..=radius).map(|x| (-(x.pow(2) as f32) / (2.0 * sigma.powi(2))).exp()),
     );
 
     // Return normalized kernels
@@ -38,16 +38,16 @@ pub fn gaussian_blur_kernel_1d(sigma: f64) -> (Array2<f64>, Array2<f64>) {
     (kernel_x, kernel_y)
 }
 
-pub fn gaussian_blur_kernel_2d(sigma: f64) -> Array2<f64> {
+pub fn gaussian_blur_kernel_2d(sigma: f32) -> Array2<f32> {
     assert!(sigma > 0.0, "--sigma should be > 0.0");
 
     // Generate an N×N Gaussian kernel
-    let radius = sigma.ceil() as i64 * 3;
+    let radius = sigma.ceil() as i32 * 3;
     let kernel = Array::from_shape_fn(
         (radius as usize * 2 + 1, radius as usize * 2 + 1),
         |(i, j)| {
-            let i = (i as i64 - radius) as f64;
-            let j = (j as i64 - radius) as f64;
+            let i = (i as i32 - radius) as f32;
+            let j = (j as i32 - radius) as f32;
 
             (-(j.powi(2) + i.powi(2)) / (2.0 * sigma.powi(2))).exp()
         },
@@ -57,7 +57,7 @@ pub fn gaussian_blur_kernel_2d(sigma: f64) -> Array2<f64> {
     &kernel / kernel.sum()
 }
 
-pub fn sobel_2d() -> (Array2<f64>, Array2<f64>) {
+pub fn sobel_2d() -> (Array2<f32>, Array2<f32>) {
     #[rustfmt::skip]
     let kernel_x = array![
         [-1.0, 0.0, 1.0],
@@ -106,18 +106,18 @@ mod tests {
 
     #[test]
     fn valid_gaussian_1d() {
-        let expect: [[f64; 7]; 1] = [[
+        let expect: [[f32; 7]; 1] = [[
             0.00081721, 0.02804152, 0.23392642, 0.47442967, 0.23392642, 0.02804152, 0.00081721,
         ]];
 
         for ((i, j), result) in gaussian_blur_kernel_1d(0.84089642).0.indexed_iter() {
-            assert_relative_eq!(expect[i][j], result, epsilon = 1e-8f64);
+            assert_relative_eq!(expect[i][j], result, epsilon = 1e-8f32);
         }
     }
 
     #[test]
     fn valid_gaussian_2d() {
-        let expect: [[f64; 7]; 7] = [
+        let expect: [[f32; 7]; 7] = [
             [
                 0.00000066, 0.00002291, 0.00019116, 0.00038771, 0.00019116, 0.00002291, 0.00000066,
             ],
@@ -142,7 +142,7 @@ mod tests {
         ];
 
         for ((i, j), result) in gaussian_blur_kernel_2d(0.84089642).indexed_iter() {
-            assert_relative_eq!(expect[i][j], result, epsilon = 1e-8f64);
+            assert_relative_eq!(expect[i][j], result, epsilon = 1e-8f32);
         }
     }
 
@@ -151,13 +151,13 @@ mod tests {
         for radius in 1..10 {
             let (kernel_x, kernel_y) = box_blur_kernel_1d(radius);
 
-            let expect = kernel_x.sum() / kernel_x.len() as f64;
+            let expect = kernel_x.sum() / kernel_x.len() as f32;
 
             for result in kernel_x.iter() {
                 assert_relative_eq!(expect, result);
             }
 
-            let expect = kernel_y.sum() / kernel_y.len() as f64;
+            let expect = kernel_y.sum() / kernel_y.len() as f32;
 
             for result in kernel_y.iter() {
                 assert_relative_eq!(expect, result);
@@ -170,7 +170,7 @@ mod tests {
         for radius in 1..10 {
             let kernel = box_blur_kernel_2d(radius);
 
-            let expect = kernel.sum() / kernel.len() as f64;
+            let expect = kernel.sum() / kernel.len() as f32;
 
             for result in kernel.iter() {
                 assert_relative_eq!(expect, result);
@@ -181,14 +181,14 @@ mod tests {
     #[test]
     fn valid_sobel2d() {
         #[rustfmt::skip]
-        let expect_x: [[f64; 3]; 3] = [
+        let expect_x: [[f32; 3]; 3] = [
             [-1.0, 0.0, 1.0],
             [-2.0, 0.0, 2.0],
             [-1.0, 0.0, 1.0]
         ];
 
         #[rustfmt::skip]
-        let expect_y: [[f64; 3]; 3] = [
+        let expect_y: [[f32; 3]; 3] = [
             [1.0, 2.0, 1.0],
             [0.0, 0.0, 0.0],
             [-1.0, -2.0, -1.0],
